Create swagger_product_order_management_tmf622
The United Way Product Ordering Management API provides a standard mechanism, based on TM Forum Open API, for placing an order with all of the necessary order parameters. The API consists on a simple set of operations that interact with CRM/Order.
This API allows Operators to place orders on the Orange Wholesale ﻿catalog, which includes characteristics such as pricing, product options and market.