Create party-management_tmf632
The United Way Party Management TMF 632 API provides a standard mechanism, based on TM Forum Open API, for party management. The API consists of a simple set of operations to interact with OWF party repository. Party can be:
an individual : each operator has the possibility to manage his data base of individuals such as operator interlocutor, delivery operator)
an organisation : possibility to search an organisation among the list of the market operators
Parties will be used in the other APIs to assign a role to an individual or organization in the context of the API.