Create product-inventory-management_tmf637
The United Way Product Catalog Management API TMF 637 consists of a simple set of operations that interact with inventory systems in a consistent manner. A product is created and modified consequently to a product order.

This API offer to retrieve a product or a collection of products depending on a filtrer criteria.

This is one of United Way's APIs.