Create quote-management_tmf648
The United Way Quote Management API TMF 648 provides a standardized mechanism for placing a quote with all of the necessary quote parameters. The API consists of a simple set of operations that interact with Orange Wholesale Order Negotiation systems in a consistent manner. A quote is created based on a product offer that is defined in the Orange Wholesale catalog. The quote identifies the product or set of products that are available, and includes characteristics such as pricing, product options and agreement.


This is one of United Way's APIs.